requirements.txt
python-dotenv==1.0.0
Flask==2.2.2
pymongo==4.3.3
firebase-admin==6.1.0